requirements.txt
langchain
streamlit
groq
python-dotenv
langchain_groq
nltk
scikit-learn
psutil
pandas